+++ example/lib/conf.dart
@@ -1,5 +1,6 @@
 Map<String, String> servermap = {
   'janus_ws': 'wss://janus1.januscaler.com/janus/ws',
+  'janus_ws_meetecho': 'wss://janus.conf.meetecho.com/ws',
   'janus_rest': 'https://janus.conf.meetecho.com/janus',
   'servercheap': 'ws://107.152.35.248:8188'
 };